Make cosine similarity faster through normalization

+++ server/src/test/java/org/elasticsearch/cluster/ClusterStateTests.java
@@ -1153,7 +1153,7 @@ public class ClusterStateTests extends ESTestCase {
     public static int expectedChunkCount(ToXContent.Params params, ClusterState clusterState) {
         final var metrics = ClusterState.Metric.parseString(params.param("metric", "_all"), true);
 
-        int chunkCount = 0;
+        long chunkCount = 0;
 
         // header chunk
         chunkCount += 1;
@@ -1217,6 +1217,6 @@ public class ClusterStateTests extends ESTestCase {
             }
         }
 
-        return chunkCount;
+        return (int) chunkCount;
     }
 }
